Move ComponentTimeout to internal package
