--- src/main/kotlin/io/github/freya022/botcommands/api/components/data/ComponentTimeout.kt
@@ -1,3 +1,0 @@
-package io.github.freya022.botcommands.api.components.data
-
-interface ComponentTimeout
