+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/builder/EphemeralTimeoutableComponentImpl.kt
@@ -2,7 +2,7 @@ package io.github.freya022.botcommands.internal.components.builder
 
 import io.github.freya022.botcommands.api.components.Components
 import io.github.freya022.botcommands.api.components.builder.IEphemeralTimeoutableComponent
-import io.github.freya022.botcommands.internal.components.data.EphemeralTimeout
+import io.github.freya022.botcommands.internal.components.data.timeout.EphemeralTimeout
 import io.github.freya022.botcommands.internal.utils.takeIfFinite
 import io.github.freya022.botcommands.internal.utils.throwArgument
 import kotlin.time.Duration
