+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/data/ComponentData.kt
@@ -1,8 +1,8 @@
 package io.github.freya022.botcommands.internal.components.data
 
-import io.github.freya022.botcommands.api.components.data.ComponentTimeout
 import io.github.freya022.botcommands.internal.components.ComponentType
 import io.github.freya022.botcommands.internal.components.LifetimeType
+import io.github.freya022.botcommands.internal.components.data.timeout.ComponentTimeout
 import kotlinx.datetime.Instant
 import kotlin.time.Duration
