+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/data/timeout/ComponentTimeout.kt
@@ -1,0 +1,4 @@
+package io.github.freya022.botcommands.internal.components.data.timeout
+
+//TODO internal
+sealed interface ComponentTimeout
