+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/data/timeout/EphemeralTimeout.kt
@@ -1,0 +1,5 @@
+package io.github.freya022.botcommands.internal.components.data.timeout
+
+internal class EphemeralTimeout(
+    val handler: suspend () -> Unit
+) : ComponentTimeout
