+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/repositories/ComponentRepository.kt
@@ -12,7 +12,12 @@ import io.github.freya022.botcommands.api.core.service.annotations.BService
 import io.github.freya022.botcommands.internal.components.ComponentType
 import io.github.freya022.botcommands.internal.components.LifetimeType
 import io.github.freya022.botcommands.internal.components.controller.ComponentFilters
-import io.github.freya022.botcommands.internal.components.data.*
+import io.github.freya022.botcommands.internal.components.data.ComponentData
+import io.github.freya022.botcommands.internal.components.data.ComponentGroupData
+import io.github.freya022.botcommands.internal.components.data.EphemeralComponentData
+import io.github.freya022.botcommands.internal.components.data.PersistentComponentData
+import io.github.freya022.botcommands.internal.components.data.timeout.EphemeralTimeout
+import io.github.freya022.botcommands.internal.components.data.timeout.PersistentTimeout
 import io.github.freya022.botcommands.internal.components.handler.EphemeralHandler
 import io.github.freya022.botcommands.internal.components.handler.PersistentHandler
 import io.github.freya022.botcommands.internal.core.db.InternalDatabase
